refactor: share mobile filter sheet controls
